paddle2onnx minor improvement
